+++ CHANGELOG.md
@@ -6,4 +6,8 @@
 ##1.0.13
 * fixed error when pagination and records counter appeared even if [counterFieldName]  is not presented in a server's response.
 ##1.0.14
-* fixed unexpected behavior of cell with overflow
+* fixed unexpected behavior of cell with overflow.
+##1.0.15
+* added api, onBeforeRequestData and onAfterSuccessfulRequestData properties. See readme for more explanations.
+##1.0.16
+* added onAfterFailedRequestData and onAfterRequestData properties